a faire, distance wall

+++ srcs/init_map/init_map.c
@@ -6,7 +6,7 @@
 /*   By: pganglof <pganglof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/26 12:56:49 by pganglof          #+#    #+#             */
-/*   Updated: 2019/11/27 18:27:57 by pganglof         ###   ########.fr       */
+/*   Updated: 2019/11/29 19:10:11 by pganglof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ int			init_map(char *av, t_map *map, t_pos *gamer)
 	if (fill_map(fd, map, gamer) == -1)
 		return (ft_close(fd));
 	close(fd);
-	map->dis_proj_plane = (int)((PROJ_PLANE_X / 2)
+	map->dis_proj_plane = (float)((PROJ_PLANE_X / 2)
 	/ tan((ALPHA / 2) * M_PI / 180));
 	return (1);
 }
